+++ sdk/src/main/java/ly/count/android/sdk/ModuleAttribution.java
@@ -1,5 +1,6 @@
 package ly.count.android.sdk;
 
+import android.content.Intent;
 import androidx.annotation.Nullable;
 
 public class ModuleAttribution extends ModuleBase {
@@ -36,24 +37,35 @@ public class ModuleAttribution extends ModuleBase {
     }
 
     @Override
+    void initFinished(CountlyConfig config) {
+        //check if any indirect attribution value is set
+        if(config.iaAttributionId != null) {
+            if(config.iaAttributionId.isEmpty()) {
+                L.e("[ModuleAttribution] provided attribution ID for indirect attribution is empty string.");
+            } else {
+                requestQueueProvider.sendIndirectAttribution(config.iaAttributionId);
+            }
+        }
+
+        //checking if any direct attribution value is set
+        if(config.daCampaignUserId != null || config.daCampaignId != null) {
+            if(config.daCampaignId == null || config.daCampaignId.isEmpty()) {
+                L.e("[ModuleAttribution] Can't record direct attribution can't be recorded with an invalid campaign id.");
+            } else {
+                if(config.daCampaignUserId != null && config.daCampaignUserId.isEmpty()){
+                    L.e("[ModuleAttribution] For direct attribution the provided Campaign user ID can't be empty string.");
+                }
+                requestQueueProvider.sendDirectAttribution(config.daCampaignId, config.daCampaignUserId);
+            }
+        }
+    }
+
+    @Override
     public void halt() {
         attributionInterface = null;
     }
 
     public class Attribution {
-
-        /**
-         * Report user attribution manually
-         *
-         * @param campaignId
-         */
-        public void recordDirectAttribution(String campaignId) {
-            synchronized (_cly) {
-                L.i("[Attribution] calling 'recordCampaign'");
-
-                recordDirectAttributionInternal(campaignId, null);
-            }
-        }
 
         /**
          * Report direct user attribution