+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 
   * Device ID can now be changed when no consent is given
   * Push notification now display/use the sent badge number. It's visualization depends on the launcher.
-  * Added a manual way to report attribution.
+  * Added new calls to report attribution.
   * Adding an init time configuration option to set the maximum request queue size
 
   * Fixed a bug where events would be sent with a larger delay if SDK callbacks are not called